feat: use a universal ascender for all fonts.
this doesn't change much for regular fonts, but makes symbols and emojis aligned with the rest of the text.

feat: use the is_fallback flag to compute a different scaling for regular fonts and symbol fonts. symbol glyphs are always scaled to fit the box and are centered. this is better as the symbol font might not be monospaced.

feat: use an additional stroke() to render all fallback fonts.